Support for finalization functions in ThreadPool
