+++ source/PTL/TaskGroup.icc
@@ -30,6 +30,7 @@
 // ---------------------------------------------------------------
 
 #include "PTL/Types.hh"
+#include "PTL/Utility.hh"
 
 namespace PTL
 {
