+++ source/PTL/Types.hh
@@ -123,42 +123,6 @@ GetSharedPointerPairMasterInstance()
 
 //======================================================================================//
 
-struct ScopeDestructor
-{
-    template <typename FuncT>
-    ScopeDestructor(FuncT&& _func)
-    : m_functor(std::forward<FuncT>(_func))
-    {}
-
-    // delete copy operations
-    ScopeDestructor(const ScopeDestructor&) = delete;
-    ScopeDestructor& operator=(const ScopeDestructor&) = delete;
-
-    // allow move operations
-    ScopeDestructor(ScopeDestructor&& rhs) noexcept
-    : m_functor(std::move(rhs.m_functor))
-    {
-        rhs.m_functor = []() {};
-    }
-
-    ScopeDestructor& operator=(ScopeDestructor&& rhs) noexcept
-    {
-        if(this != &rhs)
-        {
-            m_functor     = std::move(rhs.m_functor);
-            rhs.m_functor = []() {};
-        }
-        return *this;
-    }
-
-    ~ScopeDestructor() { m_functor(); }
-
-private:
-    std::function<void()> m_functor = []() {};
-};
-
-//======================================================================================//
-
 }  // namespace PTL
 
 // Forward declation of void type argument for usage in direct object
